Trace And Normal Of A Matrix
Finding the trace and normal of a given matrix.Here the trace is the sum of the elements of the main diagonal I.e the diagonal from the upper left to the lower right of a matrix. Normal of the matrix is the square root of the sum of all the elements.